# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fish/voices:
    get:
      operationId: listFishVoices
      tags: [fish]
      summary: List saved voice models
      responses:
        "200":
          description: List of voices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FishVoice"
    post:
      operationId: createFishVoice
      tags: [fish]
      summary: Save a new voice model by reference ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FishVoiceInput"
      responses:
        "201":
          description: Created voice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FishVoice"
  /fish/voices/{id}:
    delete:
      operationId: deleteFishVoice
      tags: [fish]
      summary: Delete a saved voice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /fish/voices/{id}/select:
    post:
      operationId: selectFishVoice
      tags: [fish]
      summary: Set a voice as the active TTS voice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Selected voice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FishVoice"
components:
  schemas:
...
    FishVoice:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        fishReferenceId:
          type: string
        isSelected:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - fishReferenceId
        - isSelected
        - createdAt
    FishVoiceInput:
      type: object
      properties:
        name:
          type: string
        fishReferenceId:
          type: string
      required:
        - name
        - fishReferenceId