# lib/api-zod/src/generated/types/fishVoice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FishVoice {
  id: number;
  name: string;
  fishReferenceId: string;
  isSelected: boolean;
  createdAt: Date;
}